Move 'text' to private folder
